feat(server): split Title & Desc components
- Rename to "Bluefin Server" as Jorge says
- Simplify the description a lot
- Remove card styling and center "Little Bluefin" quote section
- [ ] WIP for KnuckleDemos: Add details related to "justfile-driven" &
"designed/dogfood for Bluefin maintainers/maintentance" in Highlights
- [ ] WIP for KnuckleApps: Remove/simplify "Why Bluefin Server?" & "One
node to start..." section

https://opncd.ai/share/3VN9KFJe
